mental_ai_bot/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.7
anthropic==0.7.8
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
pandas==2.1.4
numpy==1.26.4

scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
google-generativeai
email-validator
